+++ packages/client/lib/sync/fetcher/blockfetcher.ts
@@ -1,5 +1,5 @@
 import { Block, BlockBuffer } from '@ethereumjs/block'
-import { KECCAK256_RLP, KECCAK256_RLP_ARRAY, BN } from 'ethereumjs-util'
+import { KECCAK256_RLP, KECCAK256_RLP_ARRAY } from 'ethereumjs-util'
 import { Peer } from '../../net/peer'
 import { Job } from './types'
 import { BlockFetcherBase, JobTask, BlockFetcherOptions } from './blockfetcherbase'
@@ -22,9 +22,12 @@ export class BlockFetcher extends BlockFetcherBase<Block[], Block> {
    * @param job
    */
   async request(job: Job<JobTask, Block[], Block>): Promise<Block[]> {
-    const { task, peer } = job
-    const { first, count } = task
-
+    const { task, peer, partialResult } = job
+    let { first, count } = task
+    if (partialResult) {
+      first = first.addn(partialResult.length)
+      count -= partialResult.length
+    }
     const blocksRange = `${first}-${first.addn(count)}`
     const peerInfo = `id=${peer?.id.slice(0, 8)} address=${peer?.address}`
 
@@ -76,37 +79,17 @@ export class BlockFetcher extends BlockFetcherBase<Block[], Block> {
    * @returns results of processing job or undefined if job not finished
    */
   process(job: Job<JobTask, Block[], Block>, result: Block[]) {
+    result = (job.partialResult ?? []).concat(result)
+    job.partialResult = undefined
     if (result.length === job.task.count) {
       return result
-    }
-    if (result.length > 0 && result.length < job.task.count) {
+    } else if (result.length > 0 && result.length < job.task.count) {
       // Adopt the start block/header number from the remaining jobs
       // if the number of the results provided is lower than the expected count
+      job.partialResult = result
       this.debug(
         `Adopt start block/header number from remaining jobs (provided=${result.length} expected=${job.task.count})`
       )
-      const lengthDiff = job.task.count - result.length
-      const adoptedJobs = []
-      let lastTask
-      while (this.in.length > 0) {
-        const job = this.in.remove()
-        if (job) {
-          lastTask = job.task
-          job.task.first = job.task.first.subn(lengthDiff)
-          adoptedJobs.push(job)
-        }
-      }
-      for (const job of adoptedJobs) {
-        this.in.insert(job)
-      }
-      if (lastTask) {
-        const tasks = this.tasks(lastTask.first.addn(lastTask.count), new BN(lengthDiff))
-        for (const task of tasks) {
-          this.enqueueTask(task)
-        }
-      }
-
-      return result
     }
     return
   }