+++ packages/client/test/rpc/engine/newPayloadV1.spec.ts
@@ -125,7 +125,8 @@ tape(`${method}: call with valid data`, async (t) => {
 })
 
 tape(`${method}: call with valid data but invalid transactions`, async (t) => {
-  const { server } = await setupChain(genesisJSON, 'post-merge', { engine: true })
+  const { chain, server } = await setupChain(genesisJSON, 'post-merge', { engine: true })
+  chain.config.logger.silent = true
   const blockDataWithInvalidTransaction = {
     ...blockData,
     transactions: ['0x1'],
@@ -144,7 +145,8 @@ tape(`${method}: call with valid data but invalid transactions`, async (t) => {
 })
 
 tape(`${method}: call with valid data & valid transaction but not signed`, async (t) => {
-  const { server, common } = await setupChain(genesisJSON, 'post-merge', { engine: true })
+  const { server, common, chain } = await setupChain(genesisJSON, 'post-merge', { engine: true })
+  chain.config.logger.silent = true
 
   // Let's mock a non-signed transaction so execution fails
   const tx = FeeMarketEIP1559Transaction.fromTxData(